% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylavaan.R
\name{as_tbl_graph.lavaan}
\alias{as_tbl_graph.lavaan}
\title{Method for importing lavaan output to tbl_graph}
\usage{
\method{as_tbl_graph}{lavaan}(object, standardize = TRUE)
}
\arguments{
\item{object}{a lavaan object, returned from lavaan's sem/cfa functions.}

\item{standardize}{should the edge coeffciants be standerdized. curretnyl only `TRUE` is supported.}
}
\value{
a tbl_graph object that can further be processed with tidygraph, and plotted with ggraph.
}
\description{
Method for importing lavaan output to tbl_graph
}
\examples{
see README
}
\author{
Mattan S. Ben-Shachar
}
